+++ react/toyproject/src/App.jsx
@@ -5,11 +5,11 @@ import Header from './Header'; // Header 컴포넌트 import
 import Sidebar from './Sidebar'; // Sidebar 컴포넌트 import
 import Mypage from './page/Mypage'; // 예시로 만든 Home 컴포넌트 import
 import AlarmList from './page/AlarmList'; // 예시로 만든 About 컴포넌트 import
-import Home from './page/Home';
 import Navigation from './component/Nav';
 import Signup from './page/Signup';
 import MusicDetail from './page/MusicDetail';
 import LoginPage from './page/LoginPage';  // 로그인 페이지 추가합니다 (04/01)
+import Root from './router/Root';
 
 
 
@@ -26,7 +26,7 @@ function App() {
             </>
           }
         >
-          <Route index element={<Home />} /> {/* 홈 페이지 라우팅 */}
+          <Route index element={<Root />} /> {/* 홈 페이지 라우팅 */}
           <Route path="/Mypage" element={<Mypage />} /> {/* 마이페이지 라우팅 */}
           <Route path="/AlarmList" element={<AlarmList />} /> {/* 알림온거 목록 경로 */}
           <Route path="/Signup" element={<Signup />} /> {/* 회원가입페이지 경로 */}
