+++ react/toyproject/src/router/SectionArtist.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Footer from './Footer';
-import './css/Home.css';
+import './css/SectionArtist.css';
 import { useNavigate } from 'react-router-dom';
 
 function Home() {
@@ -21,9 +20,7 @@ function Home() {
 
   const sectionsData = [
     // 임시배열 컴포넌트로 바꿀 예정
-    { id: 1, title: '인기아티스트', viewAllLink: '/PopularArtists', viewAllText: '모두보기' },
-    { id: 2, title: '신규아티스트', viewAllLink: '/NewArtists', viewAllText: '모두보기' },
-    { id: 3, title: '인디아티스트', viewAllLink: '/IndieArtists', viewAllText: '모두보기' }
+    { id: 1, title: '인기아티스트', viewAllLink: '/PopularArtists'},
   ];
 
   const articlesData = [
@@ -39,7 +36,7 @@ function Home() {
             <div>
               <h2>
                 <Link to={section.viewAllLink}>{section.title}</Link>
-                <Link to={section.viewAllLink} id="aaa">{section.viewAllText}</Link>
+                <Link to={section.viewAllLink} id="aaa">모두보기</Link>
               </h2>
             </div>
           </div>
@@ -61,7 +58,7 @@ function Home() {
           ))}
         </section>
       ))}
-      <Footer />
+
     </div>
   );
 }
