+++ react/toyproject/src/router/css/SectionRadio.css
@@ -1,0 +1,113 @@
+.home-background {
+    background-color: #242424;
+   
+  }
+
+  .popular-artists {
+    margin: 20px;
+    width: 100%;
+  }
+  
+  .popular-artists h2 {
+    font-size: 24px;
+    color: #ffffff; /* 텍스트 색상을 원하는 색으로 변경하세요 */
+  }
+  
+  .popular-artists a:hover {
+    text-decoration: underline;
+  }
+  
+  #aaa {
+    float: right ;
+    color: #999999;
+  }
+  
+  img {
+    size: 20px; /* 이 부분은 CSS에서는 지원되지 않습니다. 이미지 크기를 조절하려면 width와 height를 사용하세요. */
+  }
+  
+  section {
+    display: flex;
+    flex-wrap: wrap;
+  }
+  
+  /* 카드 컨테이너 */
+  article {
+    width: 200px; /* 카드의 너비 */
+    border-radius: 10px; /* 카드 모서리 둥글게 */
+    background-color: rgba(0, 0, 0, 0); /* 배경색을 투명하게 설정 */
+    overflow: hidden; /* 자식 요소가 부모 요소를 벗어나지 않도록 설정 */
+   
+  }
+  /* 이미지에 마우스 호버 시 버튼 표시 */
+  article:hover {
+    background-color: rgba(146, 146, 146, 0.1); /* 배경색을 투명하게 설정 */
+  }
+  article:active {
+    background-color: rgba(0,0, 0, 0.1); /* 배경색을 투명하게 설정 */
+  }
+  
+  /* 이미지 스타일 */
+  
+  figure img {
+    width: 98%;
+    height: auto;
+    object-fit: cover;
+    border-radius: 50%; /* 이미지를 둥글게 만듭니다. */
+    transform: translate(1%, 1%); /* 가운데 위치고정 */
+  }
+  /* 버튼 스타일 */
+  /* article-parents 요소 */
+  .article-parents {
+    display: flex;
+    position: relative; /* 상대 위치 설정 */
+    justify-content: space-between; /* 요소들을 가로 방향으로 공간을 균등하게 배치합니다. */
+  }
+  .article-parents figure {
+    margin: 0; /* 기본적인 마진을 제거합니다. */
+    flex: 1; /* 요소들의 너비를 동일하게 설정합니다. */
+  }
+  
+  
+  /* 버튼 숨기기 */
+  figure button {
+    opacity: 0; /* 버튼 투명하게 */
+    background-color: transparent; /* 버튼 투명색 */
+    transition: opacity 0.3s, background-color 0.3s; /* 투명도와 배경색 전환 효과 통합 적용 */
+    color: black; /* 임시 검은색 버튼 텍스트 색상 <- 아이콘으로 갈아탈 예정 */
+    border: none; /* 테두리 없애기 */
+    cursor: pointer; /* 포인터 커서 */
+    position: absolute; /* 절대 위치 설정 */
+    width: 100%; /* 부모 요소(figure)의 너비에 맞게 버튼의 너비 설정 */
+    height: 100%; /* 부모 요소(figure)의 높이에 맞게 버튼의 높이 설정 */
+    top: 0; /* 부모 요소의 가장 위에 위치 */
+    left: 0; /* 부모 요소의 가장 왼쪽에 위치 */
+  }
+  /* 이미지에 마우스 호버 시 버튼 표시 */
+  figure:hover button {
+    opacity: 1; /* 버튼 불투명하게 */
+  }
+  
+  .article-parents .icon {
+    position: absolute;
+    top: 54%;
+    left: 79%;
+    transform: translate(-50%, -50%); /* 이미지의 중앙에 위치하도록 설정 */
+    border-radius: 50%; /* 동그라미 모양으로 만듭니다. */
+    width: 50px; /* 동그라미의 크기 */
+    height: 50px; /* 동그라미의 크기 */
+    background-color: rgba(53, 255, 137);/* 배경색을 투명도가 있는 검은색으로 설정 */
+    color: rgb(0, 0, 0); /* 아이콘 색상을 흰색으로 설정 */
+    text-align: center; /* 텍스트를 가운데 정렬합니다. */
+    font-size: 30px; /* 아이콘 크기 */
+  }
+  .article-parents .icon:hover {
+    background-color: rgb(21, 231, 108);/* 배경색을 투명도가 있는 검은색으로 설정 */
+  }
+  .article-parents .icon:active {
+    background-color: rgb(33, 155, 84);/* 배경색을 투명도가 있는 검은색으로 설정 */
+  }
+  figure div p,figure div h4{
+   padding-left: 10px;
+   font-size: 13px;
+  }
